scripts/publisher/requirements.txt
asyncua
paho-mqtt
dotenv